refactor: 엔티티 이름은 단수형으로 통일
